split utils to tasks and utils
