+++ nl_open_data/flows/cbs/test_zip_csv_flow.py
@@ -2,7 +2,7 @@ import inspect
 from prefect import task, Flow, unmapped, Parameter
 from prefect.tasks.shell import ShellTask
 
-import nl_open_data.utils as nlu
+import nl_open_data.tasks as nlt
 
 # from nl_open_data.utils import (
 #     curl_cmd,
@@ -16,8 +16,9 @@ import nl_open_data.utils as nlu
 # )
 
 # Converting functions to tasks
-for name, _ in inspect.getmembers(nlu, inspect.isfunction):
-    exec(f"{name} = task(nlu.{name})")
+for name, _ in inspect.getmembers(nlt, inspect.isfunction):
+    exec(f"nlt.{name} = task(nlt.{name})")
+    # exec(f"{name} = task(nlt.{name})")
 
 
 # curl_cmd = task(curl_cmd)
@@ -29,7 +30,9 @@ for name, _ in inspect.getmembers(nlu, inspect.isfunction):
 # create_dir = task(create_dir)
 # remove_dir = task(remove_dir)
 
-unzip.skip_on_upstream_skip = False
+
+nlt.unzip.skip_on_upstream_skip = False
+# unzip.skip_on_upstream_skip = False
 
 curl_download = ShellTask(name="curl_download")
 
@@ -46,22 +49,47 @@ with Flow("PC6HUISNR") as zip_flow:
     dataset_description = Parameter("dataset_description", default=None)
     source = Parameter("source", required=False)
 
-    local_dir = create_dir(local_folder)
-    curl_command = curl_cmd(url, filepath)
+    # local_dir = create_dir(local_folder)
+    # curl_command = curl_cmd(url, filepath)
+    # curl_download = curl_download(command=curl_command, upstream_tasks=[local_dir])
+    # unzipped_folder = unzip(filepath, upstream_tasks=[curl_download])
+    # csv_files = list_dir(unzipped_folder, upstream_tasks=[unzipped_folder])
+    # pq_files = csv_to_parquet.map(
+    #     csv_files, delimiter=unmapped(csv_delimiter), upstream_tasks=[csv_files]
+    # )
+    # gcs_ids = upload_to_gcs.map(
+    #     to_upload=pq_files,
+    #     gcs_folder=unmapped(gcs_folder),
+    #     config=unmapped(config),
+    #     gcp_env=unmapped(gcp_env),
+    #     upstream_tasks=[pq_files],
+    # )
+    # tables = gcs_to_bq(
+    #     gcs_folder=gcs_folder,
+    #     dataset_name=dataset_name,
+    #     config=config,
+    #     gcp_env=gcp_env,
+    #     source=source,
+    #     upstream_tasks=[gcs_ids],
+    # )
+    # remove_dir(local_dir, upstream_tasks=[gcs_ids])
+
+    local_dir = nlt.create_dir(local_folder)
+    curl_command = nlt.curl_cmd(url, filepath)
     curl_download = curl_download(command=curl_command, upstream_tasks=[local_dir])
-    unzipped_folder = unzip(filepath, upstream_tasks=[curl_download])
-    csv_files = list_dir(unzipped_folder, upstream_tasks=[unzipped_folder])
-    pq_files = csv_to_parquet.map(
+    unzipped_folder = nlt.unzip(filepath, upstream_tasks=[curl_download])
+    csv_files = nlt.list_dir(unzipped_folder, upstream_tasks=[unzipped_folder])
+    pq_files = nlt.csv_to_parquet.map(
         csv_files, delimiter=unmapped(csv_delimiter), upstream_tasks=[csv_files]
     )
-    gcs_ids = upload_to_gcs.map(
+    gcs_ids = nlt.upload_to_gcs.map(
         to_upload=pq_files,
         gcs_folder=unmapped(gcs_folder),
         config=unmapped(config),
         gcp_env=unmapped(gcp_env),
         upstream_tasks=[pq_files],
     )
-    tables = gcs_to_bq(
+    tables = nlt.gcs_to_bq(
         gcs_folder=gcs_folder,
         dataset_name=dataset_name,
         config=config,
@@ -69,7 +97,7 @@ with Flow("PC6HUISNR") as zip_flow:
         source=source,
         upstream_tasks=[gcs_ids],
     )
-    remove_dir(local_dir, upstream_tasks=[gcs_ids])
+    nlt.remove_dir(local_dir, upstream_tasks=[gcs_ids])
 
 if __name__ == "__main__":
     from nl_open_data.config import get_config
